move image stuff to View and fix enhancements
